moved util files to /common
