+++ src/interface/web/app/common/colorUtils.ts
