+++ src/interface/web/app/common/iconUtils.tsx
